DOC-3005: Change all interactive tutorials to use Instruqt
- SQL Playground
- Learn CockroachDB
- JSON Support
- Spatial Data

All interactive tutorials now open in a new tab, in full-browser mode.

Removed the Python app interactive tutorial since that tutorial emphasizes using
Serverless, which is better. And removed the 21.2 version of the SQL Playground,
in favor of one playground running the latest stable release.

Added several redirects, as necessary.