DOC-886: Make feedback widget more visible
This PR moves the per-page docs feedback widget from
the bottom of each page to below the toc and adjusts
styling accordingly. Bottom of page placement is typical
on doc sites, but I'd like to see whether the new
position elicits more responses. On pages with long tocs,
the widget is less accessible. However, on the majory
of pages, the feedback prompt is now much more visible
throughout a user's time on  page, not just when the
bottom is reached.

This PR also:

- Changes the "No" response to prompt for
more details via a link to a templated GH issue in
our docs repo.
- Removes the "Contribute" drop-down options. The options
are almost never leveraged by external users. The source
files are still in place, so we can restore this if
we hear people miss it.