Show feedback widget in 2 places
Follow-up to https://github.com/cockroachdb/docs/pull/14436

Previously, the feedback widget would show either under the
page toc or, when there is no toc, at the bottom of the page.
Now, the widget shows up in both places. When the user
interacts with one, the other is hidden so as to avoid
confusion and/or duplicate submissions.

The footer overlaps the toc, which causes a strange UX
in some cases. That can be handled separately.