fix: update Docker single-node cluster documentation
The documentation for running a single-node cluster locally with Docker
was incorrect for the following reasons:

- The `docker run` command was passing `--net=roachnet` without telling
  the user how to create the `roachnet` network. Additionally, the
  `roachnet` network is not needed when deploying a single-node cluster.
- The `docker run` command was passing `--http-addr=localhost:8080`
  which made the DB Console listen on `localhost` where users could not
  connect to it from their host machine.
- The command to connect to the cluster using the built-in SQL client
  was using the incorrect hostname/IP, leading to x509 certificate
  errors when trying to connect to the DB Console (after fixing the
  `--http-addr` issue above).

The documentation has been updated to omit unnecessary options. A new
section was also added to encourage a user to connect to the DB Console
whereas it was previously easy to miss when working through the
documentation. Lastly, the `docker log` command was updated to `docker
logs`, the correct command.