Fix definitions of "active" and "open" connections
This page was conflating the concepts of open and active connections.
They are both valuable to monitor for different reasons. With this
patch, the correct metrics are described, and "active" is defined more
clearly.

It also incorrectly described a metric as being "per second" when it is
in fact just a count.