Add .bundle/config to .gitignore (#18435)
With recent macOS and Xcode and our toolchain, you now need to run this command before `bundle install`:

```
bundle config build.cbor --with-cflags="-Wno-incompatible-function-pointer-types"
bundle config build.posix-spawn --with-cflags="-Wno-incompatible-function-pointer-types"
```

After you run that plus `bundle install`, a new file is created in the repo, `.bundle/config`. This is a local configuration file that should not make it into the repo, I guess.