Fix more explicit URLs to use Liquid link syntax (#18848)
Addresses more of DOC-10953

Summary of changes:

- Update all links in release notes to use `link vXX.Y/foo.md` syntax

- ... including fixing cockroachcloud/ links in releases/ folder

- Update all links in versioned documentation (including includes) to
  use `link {{page.version.version}}/foo.md` syntax *except* in cases
  where it was obvious that the link was explicitly intended to point to
  a previous docs version (the only remaining case is a link to some
  on-prem `systemd` stuff that AFAICT is only in v23.2 docs and linked
  to from elsewhere)

NB. This change does *not* touch any of the "filter tab" links of the
format `https://www.cockroachlabs.com/docs/foo?filter=bar`, it's not
clear if that will work from Liquid. This change is big enough already,
so will punt to a separate change.