Update transaction pipelining docs re: parallelism
Fixes DOC-8591

Previously, we stated that the cost of writes is ~`O(1)` in the number
of inserts, which is wrong.

The reality is that despite the parallelism introduced by pipelining,
there is other work that happens for each SQL statement write that does
not come "for free".

NB. These changes are ported to all supported versions v23.2+