Move schema max object to 'backwards-incompatible' (#21016)
In the v25.4.0 release notes, `sql.schema.approx_max_object_count` was
already listed under 'Key cluster settings changes', but as discussed on
DOC-14949 this is a backwards-incompatible change, since it will prevent
users of large clusters from creating schema objects after the upgrade
until they modify this setting.

Moving it to the 'Backwards-incompatible changes' section for higher
visibility to lower the chance of an unpleasant surprise for users
during upgrade.

Addresses DOC-14949